Finally. Jest works. Eslint works
