+++ .eslintrc.cjs
@@ -1,0 +1,10 @@
+module.exports = {
+  extends: ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
+  parser: "@typescript-eslint/parser",
+  plugins: ["@typescript-eslint"],
+  root: true,
+  rules: {
+    "@typescript-eslint/no-empty-interface": "warn", // this rule could make perf worse!
+  },
+  ignorePatterns: ["build/"],
+};
