+++ packages/all/package.json
@@ -2,12 +2,7 @@
   "name": "monorepolint",
   "version": "0.4.0",
   "license": "MIT",
-  "scripts": {
-    "clean": "rm -rf build dist lib node_modules *.tgz tsconfig.tsbuildinfo",
-    "compile-typescript": "tsc --build",
-    "test": "NODE_OPTIONS=--experimental-vm-modules jest --colors --passWithNoTests",
-    "test:watch": "NODE_OPTIONS=--experimental-vm-modules jest --colors --passWithNoTests --watch"
-  },
+  "scripts": {},
   "dependencies": {
     "@monorepolint/cli": "workspace:*",
     "@monorepolint/config": "workspace:*",
@@ -17,14 +12,17 @@
     "@monorepolint/utils": "workspace:*"
   },
   "devDependencies": {
+    "@jest/globals": "^29.3.1",
     "@types/jest": "^29.2.4",
+    "@typescript-eslint/eslint-plugin": "^5.45.1",
+    "@typescript-eslint/parser": "^5.45.1",
+    "eslint": "^8.29.0",
     "jest": "^29.3.1",
     "prettier": "^2.8.0",
     "ts-jest": "^29.0.3",
     "tslib": "^2.4.1",
     "tsup": "^6.5.0",
-    "typescript": "^4.9.3",
-    "@jest/globals": "^29.3.1"
+    "typescript": "^4.9.3"
   },
   "publishConfig": {
     "access": "public"
