+++ packages/config/package.json
@@ -18,6 +18,7 @@
   "scripts": {
     "clean": "rm -rf build dist lib node_modules *.tgz tsconfig.tsbuildinfo",
     "compile-typescript": "tsc --build",
+    "lint": "eslint .",
     "test": "NODE_OPTIONS=--experimental-vm-modules jest --colors --passWithNoTests",
     "test:watch": "NODE_OPTIONS=--experimental-vm-modules jest --colors --passWithNoTests --watch",
     "transpile-typescript": "tsup --config ../../tsup.config.cjs"
@@ -30,14 +31,17 @@
     "tslib": "^2.3.1"
   },
   "devDependencies": {
+    "@jest/globals": "^29.3.1",
     "@types/jest": "^29.2.4",
+    "@typescript-eslint/eslint-plugin": "^5.45.1",
+    "@typescript-eslint/parser": "^5.45.1",
+    "eslint": "^8.29.0",
     "jest": "^29.3.1",
     "prettier": "^2.8.0",
     "ts-jest": "^29.0.3",
     "tslib": "^2.4.1",
     "tsup": "^6.5.0",
-    "typescript": "^4.9.3",
-    "@jest/globals": "^29.3.1"
+    "typescript": "^4.9.3"
   },
   "publishConfig": {
     "access": "public"
