+++ packages/docs/package.json
@@ -20,13 +20,16 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@jest/globals": "^29.3.1",
     "@types/jest": "^29.2.4",
+    "@typescript-eslint/eslint-plugin": "^5.45.1",
+    "@typescript-eslint/parser": "^5.45.1",
+    "eslint": "^8.29.0",
     "jest": "^29.3.1",
     "prettier": "^2.8.0",
     "ts-jest": "^29.0.3",
     "tslib": "^2.4.1",
-    "typescript": "^4.9.3",
-    "@jest/globals": "^29.3.1"
+    "typescript": "^4.9.3"
   },
   "browserslist": {
     "production": [
