+++ packages/rules/package.json
@@ -18,6 +18,7 @@
   "scripts": {
     "clean": "rm -rf build dist lib node_modules *.tgz tsconfig.tsbuildinfo",
     "compile-typescript": "tsc --build",
+    "lint": "eslint .",
     "test": "NODE_OPTIONS=--experimental-vm-modules jest --colors --passWithNoTests",
     "test:watch": "NODE_OPTIONS=--experimental-vm-modules jest --colors --passWithNoTests --watch",
     "transpile-typescript": "tsup --config ../../tsup.config.cjs"
@@ -38,6 +39,9 @@
     "@types/jest": "^29.2.4",
     "@types/semver": "^7.3.13",
     "@types/tmp": "^0.2.3",
+    "@typescript-eslint/eslint-plugin": "^5.45.1",
+    "@typescript-eslint/parser": "^5.45.1",
+    "eslint": "^8.29.0",
     "jest": "^29.3.1",
     "prettier": "^2.8.0",
     "tmp": "^0.2.1",
